Removed some nasty bug's in 'NewGame'.
First: It was possible to save a new game when the input field was empty.

Second: It was possible to save Games with the same name. This caused in an inconsistent data base. The biggest problem - beside two identical names - was that the game could not be deleted anymore ('GameDelete' uses the name- String as an unique identifier....)

Added subtitle in 'Main' which shows total # of games and scores